refactor(ssr-runtime): split index into multiple files (#4618)
* refactor: list all IR nodes explicitly

* feat(ssr): implement iterator:* directive

* test(ssr): add test for multiple iterator blocks

* fix(ssr-for-of): use empty array as default for missing iterator

* fix(ssr): update error in for-of directive for invalid value

* refactor(ssr): move `ClassList` to own file

* refactor(ssr): move `MutationTracker` to own file

* refactor(ssr): move `LightningElement` to own file

* refactor(ssr): move render functions to own file
