+++ packages/@lwc/ssr-runtime/src/types.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2024, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: MIT
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
+ */
+
+export type Attributes = Record<string, string | null>;
